AiAgent/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0  # For async SQLite
# pyodbc>=5.0.0  # Uncomment for SQL Server

# ML & Transformers
transformers>=4.40.0
torch>=2.0.0
huggingface-hub>=0.26.0
sentencepiece>=0.2.0
accelerate>=0.26.1

# Pydantic & Config
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Auth
PyJWT>=2.8.0

# Data Processing
numpy>=1.26.0
scikit-learn>=1.5.0
pandas>=2.2.0

# Image Processing
Pillow>=10.0

# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Web Scraping
beautifulsoup4>=4.12.3
lxml>=5.1.0

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
